增加修改功能

+++ src/main/java/com/netease/zzw/shopping/controller/GoodsController.java
@@ -90,6 +90,20 @@ public class GoodsController {
         return "redirect:/user/login";
     }
 
+    @RequestMapping(value = "/goods/edit", method = RequestMethod.GET)
+    public String publicGoods(@RequestParam(value = "id") long id,
+                              Model model) {
+        UserRoleDto userRoleDto = UserUtil.getUserRoleDto();
+        if(!userRoleDto.getUserName().isEmpty()
+                && userRoleDto.getRoleName().equals(UserConst.RoleName.seller.name())) {
+            Goods goods = goodsService.getGoodsById(id);
+            model.addAttribute("goods", goods);
+            model.addAttribute("userRoleDto", userRoleDto);
+            return "/goods/edit";
+        }
+        return "redirect:/user/login";
+    }
+
     @RequestMapping(value = "/goods/publicSubmit", method = RequestMethod.POST)
     public String toPublicGoods(@RequestParam(value = "title") String name,
                                 @RequestParam(value = "summary") String summary,
@@ -98,22 +112,59 @@ public class GoodsController {
                                 @RequestParam(value = "file") String fileName,
                                 @RequestParam(value = "detail") String description,
                                 @RequestParam(value = "price") BigDecimal price) {
-        //TODO 校验
-        long publisherId = 1;
+        UserRoleDto userRoleDto = UserUtil.getUserRoleDto();
+        if(!userRoleDto.getUserName().isEmpty()
+                && userRoleDto.getRoleName().equals(UserConst.RoleName.seller.name())) {
+            long publisherId = userService.findUserByUserName(userRoleDto.getUserName()).getId();
+            String graphName;
+            if (fileName == null || fileName.isEmpty()) {
+                graphName = getFileNameAndSuffixFromPath(graphLink);
+            } else {
+                graphName = fileName;
+            }
 
-        String graphName;
-        if (fileName == null || fileName.isEmpty()) {
-            graphName = getFileNameAndSuffixFromPath(graphLink);
-        } else {
-            graphName = fileName;
+            // 验证价格是否小于0
+            if(price.compareTo(new BigDecimal(0)) < 0) {
+                return "redirect:/goods/public";
+            }
+            goodsService.addGoods(name, publisherId, price, summary, description, graphName, graphSource, graphLink);
+            return "redirect:/";
         }
 
-        // 验证价格是否小于0
-        if(price.compareTo(new BigDecimal(0)) < 0) {
-            return "/goods/public";
+        return "redirect:/user/login";
+    }
+
+    @RequestMapping(value = "/goods/editSubmit", method = RequestMethod.POST)
+    public String toPublicGoods(@RequestParam(value = "id") long id,
+                                @RequestParam(value = "title") String name,
+                                @RequestParam(value = "summary") String summary,
+                                @RequestParam(value = "pic") String graphSource,
+                                @RequestParam(value = "image") String graphLink,
+                                @RequestParam(value = "file") String fileName,
+                                @RequestParam(value = "detail") String description,
+                                @RequestParam(value = "price") BigDecimal price) {
+        UserRoleDto userRoleDto = UserUtil.getUserRoleDto();
+        if(!userRoleDto.getUserName().isEmpty()
+                && userRoleDto.getRoleName().equals(UserConst.RoleName.seller.name())) {
+
+            long publisherId = userService.findUserByUserName(userRoleDto.getUserName()).getId();
+            String graphName;
+            if (fileName == null || fileName.isEmpty()) {
+                graphName = getFileNameAndSuffixFromPath(graphLink);
+            } else {
+                graphName = fileName;
+            }
+
+            // 验证价格是否小于0
+            if(price.compareTo(new BigDecimal(0)) < 0) {
+                return "redirect:/goods/edit?id="+id;
+            }
+            deleteGraph(id);
+            goodsService.updateGoodsById(id, name, publisherId, price, summary, description, graphName, graphSource, graphLink);
+            return "redirect:/";
         }
-        goodsService.addGoods(name, publisherId, price, summary, description, graphName, graphSource, graphLink);
-        return "redirect:/";
+
+        return "redirect:/user/login";
     }
 
     private String getFileNameAndSuffixFromPath(String path) {
@@ -125,47 +176,67 @@ public class GoodsController {
     @RequestMapping(value = "/api/goods/image/upload", method = RequestMethod.POST)
     @ResponseBody
     public String imageUpload(@RequestParam(value = "file") MultipartFile file) {
-        String fileName = file.getOriginalFilename();
-        String filePath = imageSavePath + "/";
-        fileName = UUID.randomUUID() + delimiter + fileName;
-        File dest = new File(filePath + fileName);
-        if (!dest.getParentFile().exists()) {
-            dest.getParentFile().mkdirs();
-        }
-
-        try {
-            file.transferTo(dest);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
         JSONObject resultPath = new JSONObject();
-        resultPath.put("code", 200);
-        resultPath.put("result", GoodsConst.relativePath + fileName);
+        UserRoleDto userRoleDto = UserUtil.getUserRoleDto();
+        if(!userRoleDto.getUserName().isEmpty()
+                && userRoleDto.getRoleName().equals(UserConst.RoleName.seller.name())) {
+
+            String fileName = file.getOriginalFilename();
+            String filePath = imageSavePath + "/";
+            fileName = UUID.randomUUID() + delimiter + fileName;
+            File dest = new File(filePath + fileName);
+            if (!dest.getParentFile().exists()) {
+                dest.getParentFile().mkdirs();
+            }
+
+            try {
+                file.transferTo(dest);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+
+            resultPath.put("code", 200);
+            resultPath.put("result", GoodsConst.relativePath + fileName);
+            return resultPath.toJSONString();
+        }
+        resultPath.put("code", 400);
         return resultPath.toJSONString();
     }
 
-    @RequestMapping(value = "/api/goods/delete", method = RequestMethod.POST)
-    @ResponseBody
-    public String deleteGoods(@RequestParam(value = "id") long id) {
-        JSONObject resultId = new JSONObject();
+    private void deleteGraph(long goodsId) {
         // 判断是否需要删除图片
-        Goods goods = goodsService.getGoodsById(id);
-        if(goods.getGraphSource().equals(GoodsConst.GraphSource.file.name())) {
+        Goods goods = goodsService.getGoodsById(goodsId);
+        if (goods.getGraphSource().equals(GoodsConst.GraphSource.file.name())) {
             String graphNameAndSuffix = getFileNameAndSuffixFromPath(goods.getGraphLink());
             String graphPath = imageSavePath + "/" + graphNameAndSuffix;
             File graphFile = new File(graphPath);
             // TODO 返回值是boolean，最好做一下判断看是否成功删除
             graphFile.delete();
         }
+    }
 
-        if(goodsService.deleteGoodsById(id) == 0) {
-            resultId.put("code", 404);
-            resultId.put("id", -1);
+    @RequestMapping(value = "/api/goods/delete", method = RequestMethod.POST)
+    @ResponseBody
+    public String deleteGoods(@RequestParam(value = "id") long id) {
+        JSONObject resultId = new JSONObject();
+
+        UserRoleDto userRoleDto = UserUtil.getUserRoleDto();
+        if(!userRoleDto.getUserName().isEmpty()
+                && userRoleDto.getRoleName().equals(UserConst.RoleName.seller.name())) {
+
+            deleteGraph(id);
+            if (goodsService.deleteGoodsById(id) == 0) {
+                resultId.put("code", 404);
+                resultId.put("id", -1);
+                return resultId.toJSONString();
+            }
+            resultId.put("code", 200);
+            resultId.put("id", id);
             return resultId.toJSONString();
         }
-        resultId.put("code", 200);
-        resultId.put("id", id);
+
+        resultId.put("code", 400);
         return resultId.toJSONString();
     }
 }
